requirements.txt
pre-commit
djangorestframework
djangorestframework-simplejwt
drf_spectacular